refactor to remove defprotocol ForOpenAIApi - moved the code and tests into the ChatOpenAI specific modules - updated references in other models
