+++ lib/chat_models/chat_google_ai.ex
@@ -10,10 +10,10 @@ defmodule LangChain.ChatModels.ChatGoogleAI do
   alias __MODULE__
   alias LangChain.Config
   alias LangChain.ChatModels.ChatModel
+  alias LangChain.ChatModels.ChatOpenAI
   alias LangChain.Message
   alias LangChain.MessageDelta
   alias LangChain.LangChainError
-  alias LangChain.ForOpenAIApi
   alias LangChain.Utils
 
   @behaviour ChatModel
@@ -51,7 +51,7 @@ defmodule LangChain.ChatModels.ChatGoogleAI do
     # 1 means the selected token is the most probable among all the tokens in the
     # model's vocabulary (also called greedy decoding), while a topK of 3 means that
     # the next token is selected from among the 3 most probable using the temperature.
-    # For each token selection step, the topK tokens with the highest probabilities 
+    # For each token selection step, the topK tokens with the highest probabilities
     # are sampled. Tokens are then further filtered based on topP with the final token
     # selected using temperature sampling.
     field :top_k, :float, default: 1.0
@@ -140,7 +140,7 @@ defmodule LangChain.ChatModels.ChatGoogleAI do
         %{
           # Google AI functions use an OpenAI compatible format.
           # See: https://ai.google.dev/docs/function_calling#how_it_works
-          "functionDeclarations" => Enum.map(functions, &ForOpenAIApi.for_api/1)
+          "functionDeclarations" => Enum.map(functions, &ChatOpenAI.for_api/1)
         }
       ])
     else
